+++ src/main/java/be/bonamis/advent/utils/marsrover/Position.java
@@ -1,4 +1,4 @@
 package be.bonamis.advent.utils.marsrover;
 
-public record Position(int x, int y) {
+public record Position(long x, long y) {
 }
